# Collected HGCal reconstruction changes