--- a/RecoLocalCalo/HGCalRecAlgos/interface/HGCalDepthPreClusterer.h
+++ b/RecoLocalCalo/HGCalRecAlgos/interface/HGCalDepthPreClusterer.h
@@ -14,12 +14,12 @@ class HGCalDepthPreClusterer
 {
 public:
   
- HGCalDepthPreClusterer() : radius(0.), minClusters(0.), realSpaceCone(false), clusterTools(nullptr)
+ HGCalDepthPreClusterer() : radii({0.,0.,0.,}), minClusters(0.), realSpaceCone(false), clusterTools(nullptr)
     {
   }
   
- HGCalDepthPreClusterer(const edm::ParameterSet& conf, edm::ConsumesCollector& sumes, double radius_in, uint32_t min_clusters, bool real_space_cone) : 
-  radius(radius_in),
+ HGCalDepthPreClusterer(const edm::ParameterSet& conf, edm::ConsumesCollector& sumes, std::vector<double> radii_in, uint32_t min_clusters, bool real_space_cone) : 
+  radii(radii_in),
   minClusters(min_clusters),
   realSpaceCone(real_space_cone),
   clusterTools(std::make_unique<hgcal::ClusterTools>(conf,sumes)) {
@@ -34,7 +34,7 @@ public:
   std::vector<reco::HGCalMultiCluster> makePreClusters(const reco::HGCalMultiCluster::ClusterCollection &) const;    
 
 private:  
-  float radius;
+  std::vector<double> radii;
   uint32_t minClusters;
   bool realSpaceCone;