--- a/RecoLocalCalo/HGCalRecAlgos/interface/HGCalImagingAlgo.h
+++ b/RecoLocalCalo/HGCalRecAlgos/interface/HGCalImagingAlgo.h
@@ -52,21 +52,11 @@ class HGCalImagingAlgo
 		      algoId(reco::CaloCluster::undefined),
 		      verbosity(pERROR){
  }
-  
-  HGCalImagingAlgo(float delta_c_in, double kappa_in, double ecut_in,
-		   //		   const CaloSubdetectorTopology *thetopology_p,
-		   reco::CaloCluster::AlgoId algoId_in,
-		   VerbosityLevel the_verbosity = pERROR) : delta_c(delta_c_in), kappa(kappa_in), 
-							    ecut(ecut_in),    
-							    cluster_offset(0),
-							    sigma2(1.0),
-							    algoId(algoId_in),
-							    verbosity(the_verbosity){
-  }
 
   HGCalImagingAlgo(float delta_c_in, double kappa_in, double ecut_in,
 		   //		   const CaloSubdetectorTopology *thetopology_p,
 		   reco::CaloCluster::AlgoId algoId_in,
+                   bool dependSensor_in,
                    std::vector<double> dEdXweights_in,
                    std::vector<double> thicknessCorrection_in,
                    std::vector<double> fcPerMip_in,
@@ -78,6 +68,7 @@ class HGCalImagingAlgo
 							    cluster_offset(0),
 							    sigma2(1.0),
 							    algoId(algoId_in),
+                                                            dependSensor(dependSensor_in),
 							    dEdXweights(dEdXweights_in),
                                                             thicknessCorrection(thicknessCorrection_in),
                                                             fcPerMip(fcPerMip_in),
@@ -86,23 +77,12 @@ class HGCalImagingAlgo
                                                             noiseMip(noiseMip_in),
 							    verbosity(the_verbosity){
   }
-  
-  HGCalImagingAlgo(float delta_c_in, double kappa_in, double ecut_in,
-		   double showerSigma, 
-		   //		   const CaloSubdetectorTopology *thetopology_p,
-		   reco::CaloCluster::AlgoId algoId_in,
-		   VerbosityLevel the_verbosity = pERROR) : delta_c(delta_c_in), kappa(kappa_in), 
-							    ecut(ecut_in),    
-							    cluster_offset(0),
-							    sigma2(std::pow(showerSigma,2.0)),
-							    algoId(algoId_in),
-							    verbosity(the_verbosity){
-  }
 
   HGCalImagingAlgo(float delta_c_in, double kappa_in, double ecut_in,
 		   double showerSigma, 
 		   //		   const CaloSubdetectorTopology *thetopology_p,
 		   reco::CaloCluster::AlgoId algoId_in,
+                   bool dependSensor_in,
                    std::vector<double> dEdXweights_in,
                    std::vector<double> thicknessCorrection_in,
                    std::vector<double> fcPerMip_in,
@@ -114,6 +94,7 @@ class HGCalImagingAlgo
 							    cluster_offset(0),
 							    sigma2(std::pow(showerSigma,2.0)),
 							    algoId(algoId_in),
+                                                            dependSensor(dependSensor_in),
 							    dEdXweights(dEdXweights_in),
                                                             thicknessCorrection(thicknessCorrection_in),
                                                             fcPerMip(fcPerMip_in),
@@ -174,7 +155,8 @@ class HGCalImagingAlgo
   // The algo id
   reco::CaloCluster::AlgoId algoId;
 
-  // various parameters used for calculating the noise levels for a given sensor
+  // various parameters used for calculating the noise levels for a given sensor (and whether to use them)
+  bool dependSensor;
   std::vector<double> dEdXweights;
   std::vector<double> thicknessCorrection;
   std::vector<double> fcPerMip;