… RecoLocalCalo/HGCalRecAlgos/interface/RecHitTools.h
… RecoLocalCalo/HGCalRecAlgos/src/ClusterTools.cc
… RecoLocalCalo/HGCalRecAlgos/src/HGCal3DClustering.cc
… RecoLocalCalo/HGCalRecAlgos/src/HGCalDepthPreClusterer.cc
… RecoLocalCalo/HGCalRecAlgos/src/HGCalImagingAlgo.cc
… RecoLocalCalo/HGCalRecAlgos/src/RecHitTools.cc
… RecoLocalCalo/HGCalRecProducers/plugins/HGCalClusterProducer.cc
… RecoLocalCalo/HGCalRecProducers/plugins/HGCalClusterTestProducer.cc
… RecoLocalCalo/HGCalRecProducers/python/HGCRecHitCollections_cfi.py
… RecoLocalCalo/HGCalRecProducers/python/hgcalLayerClusters_cfi.py