- multiclustering in Cartesian space
- allow for sensor-dependent E and rho in 2D clustering
- implementation of 3D clustering based on kdtree
- more flexible 3D clustering
- shower shape calculations