@cmsbuild 2017-04-06
A new Pull Request was created by  @clelange (Clemens Lange) for master.

It involves the following packages:

RecoLocalCalo/HGCalRecAlgos
RecoLocalCalo/HGCalRecProducers


@perrotta, @cmsbuild, @slava77, @kpedro88, @davidlange6 can you please review it and eventually sign? Thanks.
@vandreev11, @sethzenz, @kpedro88, @argiro, @cseez, @pfs, @lgray this is something you requested to watch as well.
@muzaffar, @davidlange6, @smuzaffar you are the release manager for this.

cms-bot commands are listed here cms-sw/cmssw#13028