@slava77 2017-04-06
@clelange please provide a link to slides describing changes in performance
and also please provide some timing cost estimates in PU200
Thank you.