@clelange 2017-04-06
@beaudett and @edjtscott should also watch this

We'll try to come up with a summary soon.