@cmsbuild 2017-04-07
+1
Tested at: 5d9618d6cf53dfa0dfb5f836898d212738d043cf
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-18236/18989/summary.html