@kpedro88 2017-04-07
My code review is complete: mostly just cleanup, hygiene, and safety checks. @slava77 or @perrotta may have other comments.

@lgray suggests that this PR will actually speed up HGCal reco by making more use of kd-trees.