@cmsbuild 2017-04-07
Comparison is ready
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-18236/18989/summary.html

Comparison Summary:
 - No significant changes to the logs found
 - Reco comparison results: 2420 differences found in the comparisons
 - DQMHistoTests: Total files compared: 23
 - DQMHistoTests: Total histograms compared: 1917345
 - DQMHistoTests: Total failures: 36430
 - DQMHistoTests: Total nulls: 0
 - DQMHistoTests: Total successes: 1880742
 - DQMHistoTests: Total skipped: 173
 - DQMHistoTests: Total Missing objects: 0
 - Checked 94 log files, 14 edm output root files, 23 DQM output files