@clelange 2017-04-18
Thanks for all these suggestions, I think I should have implemented them all and will push the code later today (after HGCal collaborators had the chance to have a look). I'll also provide the timing tests.
The overall summary of the code changes are:

Speedup for multiclustering algorithm (association of 2d layer
clusters). New classes using k-d trees to search for 2d clusters. This
results in a significant (factor ~1000) speedup in high PU - see
[1]. Search radius definition (in cm) replaces the previous deltaR (in eta-phi) definition. The new algorithm with k-d trees reproduces the results obtained with the older implementation.

The 2d clustering parameters E_c (a energy threshold for rechits) and
rho_c (energy density threshold for new cluster) are now defined in
terms of the local sigma_noise. Values are set at three and nine sigma
respectively. This change should also result in a small speedup to the 2D clustering.

Other small changes permit various parameters to vary by sub-detector
(EE, FH, BH), instead of being uniform throughout.

[1] https://indico.cern.ch/event/611352/contributions/2465081/attachments/1409754/2155764/hgcPU08fev17.pdf