@clelange 2017-04-18
Sorry I missed a few comments, now they should be addressed apart from the one with the magic numbers to be defined in a single place only (we can do that, but it's more convenient to do that from a user account instead of using the CMS-HGCAL team).

The results of the timing tests (in `CMSSW_9_1_0_pre1`) running on `/RelValZTT_14TeV/CMSSW_9_0_0_pre4-PU25ns_90X_upgrade2023_realistic_v3_D4TPU200c2-v1/GEN-SIM-REC`O can be found at:

https://clangeplots.web.cern.ch/clangeplots/clusteringOld.log vs.
https://clangeplots.web.cern.ch/clangeplots/clusteringNew.log

I hope that's all you need. Thanks again!