@kpedro88 2017-04-18
@clelange I'm looking through the logs you posted. It seems like the 900pre4 relval you chose doesn't run well in 910pre1 (3/4 events fail with exceptions).

old:
```
TimeReport ---------- Event  Summary ---[sec]----
TimeReport       event loop CPU/event = 187.323324
TimeReport      event loop Real/event = 104.310570
TimeReport     sum Streams Real/event = 206.931125
TimeReport efficiency CPU/Real/thread = 0.448956
--------------------------------------------------
TimeReport   3.452880    13.811519    13.811519  hgcalLayerClusters
TimeReport   0.555371     0.555371     0.555371  particleFlowClusterHGCal
TimeReport   0.597833     0.597833     0.597833  particleFlowSuperClusterHGCal
```

new:
```
TimeReport ---------- Event  Summary ---[sec]----
TimeReport       event loop CPU/event = 184.785696
TimeReport      event loop Real/event = 102.315058
TimeReport     sum Streams Real/event = 200.910353
TimeReport efficiency CPU/Real/thread = 0.451511
--------------------------------------------------
TimeReport   3.720647    14.882589    14.882589  hgcalLayerClusters
TimeReport   0.528324     0.528324     0.528324  particleFlowClusterHGCal
TimeReport   0.497209     0.497209     0.497209  particleFlowSuperClusterHGCal
```

So `hgcalLayerClusters` is a bit slower, but `particleFlow*HGCal` is a bit faster. Overall, at least the CPU usage has stayed roughly the same, but I was led to expect more of a speedup. It would be better to run igprof and process more (successful) events...