@clelange 2017-04-18
@kpedro88 OK, I will try to process more events tomorrow, maybe the more recent RelVals work better. I'll also check if we can find why there's no apparent speed-up.