@cmsbuild 2017-04-18
+1
Tested at: 38018c2a9cc399851fcba918972989bf05f56bfa
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-18236/19228/summary.html

The following merge commits were also included on top of IB + this PR after doing git cms-merge-topic: 
https://github.com/cms-sw/cmssw/commit/38018c2a9cc399851fcba918972989bf05f56bfa
https://github.com/cms-sw/cmssw/commit/c881803bca952400af31b02106302e55fbed56ea
You can see more details here:
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-18236/19228/git-log-recent-commits
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-18236/19228/git-merge-result