@cmsbuild 2017-04-19
Comparison is ready
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-18236/19228/summary.html

Comparison Summary:
 - No significant changes to the logs found
 - Reco comparison results: 2582 differences found in the comparisons
 - DQMHistoTests: Total files compared: 23
 - DQMHistoTests: Total histograms compared: 1826205
 - DQMHistoTests: Total failures: 38250
 - DQMHistoTests: Total nulls: 0
 - DQMHistoTests: Total successes: 1787782
 - DQMHistoTests: Total skipped: 173
 - DQMHistoTests: Total Missing objects: 0
 - Checked 94 log files, 14 edm output root files, 23 DQM output files