@clelange 2017-04-20
Here are the timing results based on 200 events of a ttbar RelVal sample (/RelValTTbar_14TeV/CMSSW_9_1_0_pre1-PU25ns_90X_upgrade2023_realistic_v9_D4TPU200r3-v1/GEN-SIM-RECO):

old:
```
TimeReport ---------- Event  Summary ---[sec]----
TimeReport       event loop CPU/event = 45.078342
TimeReport      event loop Real/event = 19.219300
TimeReport     sum Streams Real/event = 74.101633
TimeReport efficiency CPU/Real/thread = 0.586368

TimeReport   0.052869     0.052869     0.052869  HGCalRecHit
TimeReport   0.190864     0.190864     0.190864  HGCalUncalibRecHit
TimeReport  16.244751    16.244751    16.244751  hgcalLayerClusters
TimeReport   0.597499     0.597499     0.597499  particleFlowClusterHGCal
TimeReport   0.378978     0.378978     0.378978  particleFlowRecHitHGC
```

new:
```
TimeReport ---------- Event  Summary ---[sec]----
TimeReport       event loop CPU/event = 46.678580
TimeReport      event loop Real/event = 21.006910
TimeReport     sum Streams Real/event = 80.823090
TimeReport efficiency CPU/Real/thread = 0.555515

TimeReport   0.050563     0.050563     0.050563  HGCalRecHit
TimeReport   0.132911     0.132911     0.132911  HGCalUncalibRecHit
TimeReport  17.433431    17.433431    17.433431  hgcalLayerClusters
TimeReport   0.561105     0.561105     0.561105  particleFlowClusterHGCal
TimeReport   0.944612     0.944612     0.944612  particleFlowRecHitHGC
```

Overall, the new version is slightly slower.
I will now run IgProf again and post a link to the results.