@clelange 2017-05-17
@perrotta yes, let me summarise the changes:
There were further bugs introduced by others that I've discovered and corrected, the major one being a threshold value (`sigmaNoise`) not being initialised correctly, creating a lot more layer clusters than intended (the default value was set to a negative number, which was then compared to be less than a small but always positive energy value resulting, which is obviously true all the time).

I have produced a large number of validation plots, some of which we'd like to add to the standard validation plots in CMSSW in a separate PR in the near future. These can be found at https://clangeplots.web.cern.ch/clangeplots/HGCal/
A brief summary can be found in Chris Seez's talk to be held this afternoon: https://indico.cern.ch/event/639253/ - the numbers and plots all make sense.

In case you'd like to look at them at detail, here's some guidance:
I've made these plots for three different samples:
- single particle gun charged pions, pT = 35 GeV, PU=0
- single particle gun photons, pT = 35 GeV, PU=0
- ttbar, PU=200
In the respective subfolders you can see some "event displays" showing the layer clusters with box size proportional to their energy. From them you can see that there are only very few noise clusters for the PU=0 samples, while a lot of activity can be observed in the PU=200 sample. You can also nicely see the different calorimeter penetration depths and shower widths.

Going further to the subfolders, different distributions are shown:
- cluster2d_N: layer cluster multiplicity, split by layer for cluster2d_N_layer* plots
- cluster2d_eta: layer cluster eta distribution, split by layer for cluster2d_eta_layer* plots
- cluster2d_nhitAll: number of cell hits per cluster, also shown split per layer
- cluster2d_nhitCore: number of cell core (i.e. not counting halo/border) hits per cluster, also shown split per layer
- cluster2d_pt: layer cluster pt distribution, also shown split per layer
- multiclus_*: multiplicity (N), pt, and eta distributions for multiclusters
- rechits_N: rechit multiplicity

The RecHits do not need to be associated to a layer cluster, neither do the layer clusters have to be associated to a multicluster.

The subfolders have the following cuts applied:
- nocuts: no cuts
- nocuts_thickness*: count only sensors of 100, 200, 300 µm thickness and "other" (scintillator, i.e. BH)
- plus/minus_z: restrict to positive/negative endcap only
- plus/minus_z_etaUnit: restrict to positive/negative endcap only and 1.6 < |eta| <= 2.6

I've also performed the standard tests that were requested previously on a PU200 ttbar sample:

Timing:
old (copying from above):
```
TimeReport   3.452880    13.811519    13.811519  hgcalLayerClusters
TimeReport   0.555371     0.555371     0.555371  particleFlowClusterHGCal

new:
```
TimeReport  14.247040    14.247040    14.247040  hgcalLayerClusters
TimeReport   0.629979     0.629979     0.629979  particleFlowClusterHGCal

```

The slight increase in processing time prevails, but is expected since the thresholds are lower in average, but are chosen in a better way (see above), leading to more hits that are being considered.

EventSize (PU200 ttbar):
old:
```
recoCaloClusters_hgcalLayerClusters__RECO. 5.77026e+06 1.2683e+06
recoHGCalMultiClusters_hgcalLayerClusters__RECO. 28026.9 7699.58
```
new:
```
recoCaloClusters_hgcalLayerClusters__RECO2. 1.4619e+07 2.69955e+06
recoHGCalMultiClusters_hgcalLayerClusters__RECO2. 1.88944e+06 516619
```
As also discussed above, the number of layer clusters and multiclusters is expected to grow, but it's within reason and you can see from the validation plots and the previous discussion that this is desired and needed.